Move couch view service to libs
