+++ test/libs/couch/view.spec.ts
@@ -4,7 +4,7 @@ import { expect } from 'chai';
 import sinon, { SinonSpy, SinonStub } from 'sinon';
 import { ChtClientService } from '../../../src/services/cht-client';
 import { HttpClientRequest } from '@effect/platform';
-import { warmView } from '../../../src/services/couch/view';
+import { warmView } from '../../../src/libs/couch/view';
 import { genWithLayer, sandbox } from '../../utils/base';
 
 const FAKE_CLIENT_REQUEST = { hello: 'world' } as const;
@@ -15,7 +15,7 @@ const run = Layer
   .succeed(ChtClientService, { request: couchRequest } as unknown as ChtClientService)
   .pipe(genWithLayer);
 
-describe('Couch View Service', () => {
+describe('Couch View lib', () => {
   let requestGet: SinonSpy;
   let requestSetUrlParam: SinonStub;
